+++ lib/src/wistia_player_value.dart
@@ -330,10 +330,4 @@ class WistiaPlayerValue {
     data['wmode'] = wmode;
     return data;
   }
-
-  @override
-  String toString() {
-    List<String> values = ['autoPlay=$autoPlay'];
-    return values.join(' ');
-  }
 }