+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * Add more default value for `WistiaPlayerValue`
+
+## 0.0.3
+
+* Refactor options passing to be passed to the Wistia div class
+* Add `convertUrlToId` to convert Wistia shareable URL to Wistia ID to pass to `WistiaPlayerController` videoId property